Fix GetLastError (#35)
qdb_get_last_error was updated to use **qdb_string_t for the message instead of *qdb_string_t. I changed the message to *qdb_string_t and made the function returns an empty string if the message pointer isn't updated. In addition, I handle the memory release of the message object.