QDB-16147 - Make go api compatible with docker builder (#44)
This commit refactors the teamcity build process, and uses a similar approach like other repositories where we invoke scripts inside inside the scripts/teamcity/ directory.

This has the advantage that, instead of relying on a lot of teamcity-specific stuff, we're able to probe the kind of environment we're in, auto-detect default locations, and set variables accordingly.