QDB16359 - Add latest batch writer and bulk reader (#85)
This PR is an accumulation of a lot of work.

It adds:

* Support for the high-performance batch writer API;
* Support for the high-performance bulk reader API.

Furthermore, it does a lot of cleanup and tries to align the APIs with more idiomatic Go usage.

Uses zero-copy / memory pinning whenever possible when going from Go -> C, and uses QuasarDB's native qdb_alloc_buffer() for anything else.